Rename DefaultLogger DefaultDBLogger
