--- wallet/db/defaultlogger/defaultlogger.cpp
@@ -1,8 +1,0 @@
-#include "defaultlogger.h"
-
-#include "util.h"
-
-void DefaultLogger::logWrite(const StringViewT message)
-{
-    OutputDebugStringF("%s\n", message.to_string().c_str());
-}
